Create alert dialog widget and add confirmation dialog when delete post or comments

+++ lib/widgets/post/comment_row.dart
@@ -2,6 +2,7 @@ import 'package:discover/models/comments/comment.dart';
 import 'package:discover/utils/api/api.dart';
 import 'package:discover/utils/functions.dart';
 import 'package:discover/utils/providers/preferences_provider.dart';
+import 'package:discover/widgets/ui/custom_alert_dialog.dart';
 import 'package:discover/widgets/user/user_image.dart';
 import 'package:flutter/material.dart';
 
@@ -13,8 +14,19 @@ class CommentRow extends StatelessWidget {
       : super(key: key);
 
   Future<void> _deleteComment(BuildContext context) async {
-    // TODO: Display confirm dialog
-    // TODO: Call API to remove comment
+    bool isOk = await showDialog(
+      context: context,
+      builder: (dialogCtx) {
+        return CustomAlertDialog(
+          title: "Delete comment",
+          content: Text("Are you sure you want to delete this comment ?"),
+          onNegative: () => Navigator.of(context).pop(false),
+          onPositive: () => Navigator.of(context).pop(true),
+        );
+      },
+    );
+    if (isOk != true) return;
+
     try {
       final token = PreferencesProvider.of(context).getUser()?.token;
       await Api().deleteComment(comment.idPost, comment.id, token);
